% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filternarmed.R
\name{filternarmed}
\alias{filternarmed}
\title{Create an API for trained ensemble model}
\usage{
filternarmed(v0, x, lambda, nArms)
}
\arguments{
\item{v0}{v0 value}

\item{x}{x value}

\item{lambda}{labda value}

\item{nArms}{nArms value}
}
\value{
tempevmat returns variable
}
\description{
Create an API for trained ensemble model
}
\examples{
# filternarmed(v0,x,lambda,nArms)
}
