% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filterrlcpredpfun.R
\name{Filterrlcpredpfun}
\alias{Filterrlcpredpfun}
\title{This function computes predicted prob. matrix of EVL model}
\usage{
Filterrlcpredpfun(temppars, deckchoices, wins, losses)
}
\arguments{
\item{temppars}{temppar values}

\item{deckchoices}{deckchoice values}

\item{wins}{win values}

\item{losses}{loss values}
}
\value{
returns ppmat
}
\description{
This function computes predicted prob. matrix of EVL model
}
\examples{
# Filterrlcpredpfun(temppars, deckchoices, wins, losses)
}
