% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdmtkExprLiveStats.R
\name{rdmtkExprLiveStats}
\alias{rdmtkExprLiveStats}
\title{Generate live stats}
\usage{
rdmtkExprLiveStats(taskType, exprAID, exprBID, exprReln, riskLvl,
  sig_level, ...)
}
\arguments{
\item{taskType}{task type}

\item{exprAID}{exprAID values}

\item{exprBID}{exprBID values}

\item{exprReln}{exprReln values}

\item{riskLvl}{risk level values}

\item{sig_level}{significance level}
}
\value{
returns a list
}
\description{
Generate live stats
}
\examples{
# rdmtkExprLiveStats(taskType, exprAID, exprBID, exprReln, riskLvl, sig_level)
}
