% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdmtkGetIGTData.R
\name{rdmtkGetIGTData}
\alias{rdmtkGetIGTData}
\title{rdmtkGetIGTData}
\usage{
rdmtkGetIGTData(taskType, exprID, ...)
}
\arguments{
\item{taskType}{task type}

\item{exprID}{exprAID values}
}
\value{
returns a returnIGTExprData
}
\description{
rdmtkGetIGTData
}
\examples{
# rrdmtkGetIGTData(taskType, exprID)
}
