% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdmtkGetParticipantsCntExpr.R
\name{rdmtkGetParticipantsCntExpr}
\alias{rdmtkGetParticipantsCntExpr}
\title{Get participant counts}
\usage{
rdmtkGetParticipantsCntExpr(taskType, exprAID, exprBID, ...)
}
\arguments{
\item{taskType}{task type}

\item{exprAID}{exprAID values}

\item{exprBID}{exprBID values}
}
\value{
100
}
\description{
Get participant counts
}
\examples{
# rdmtkGetParticipantsCntExpr(taskType, exprAID, exprBID)
}
