% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdmtkIGTBaselineModel.R
\name{rdmtkIGTBaselineModel}
\alias{rdmtkIGTBaselineModel}
\title{This program does the EVL, Baseline, & Random Models for the IGT}
\usage{
rdmtkIGTBaselineModel(taskType, exprID, ...)
}
\arguments{
\item{taskType}{task type}

\item{exprID}{exprAID values}
}
\value{
100
}
\description{
This program does the EVL, Baseline, & Random Models for the IGT
}
\examples{
# rdmtkIGTBaselineModel(taskType, exprID)
}
