% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdmtkIGTEVLModel.R
\name{rdmtkIGTEVLModel}
\alias{rdmtkIGTEVLModel}
\title{This program does the EVL (expectancy-valance learning model), Baseline, & Random Models for the IGT}
\usage{
rdmtkIGTEVLModel(taskType, exprID, ...)
}
\arguments{
\item{taskType}{task type}

\item{exprID}{exprAID values}
}
\value{
100
}
\description{
This program does the EVL (expectancy-valance learning model), Baseline, & Random Models for the IGT
}
\examples{
# rdmtkIGTBaselineModel(taskType, exprID)
}
