% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdmtkIGTMdlRslts.R
\name{rdmtkIGTMdlRslts}
\alias{rdmtkIGTMdlRslts}
\title{Generate results and input to database}
\usage{
rdmtkIGTMdlRslts(taskType, exprID, mdlType, ...)
}
\arguments{
\item{taskType}{task type}

\item{exprID}{exprAID values}

\item{mdlType}{type of model}
}
\value{
a returnList
}
\description{
Generate results and input to database
}
\examples{
# rdmtkIGTMdlRslts(taskType, exprID, mdlType)
}
