% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdmtkIGTRandomModel.R
\name{rdmtkIGTRandomModel}
\alias{rdmtkIGTRandomModel}
\title{This program does the EVL, Baseline, & Random Models for the IGT}
\usage{
rdmtkIGTRandomModel(taskType, exprID, ...)
}
\arguments{
\item{taskType}{task type}

\item{exprID}{exprAID values}
}
\value{
a returnList
}
\description{
This program does the EVL, Baseline, & Random Models for the IGT
}
\examples{
# rdmtkIGTRandomModel(taskType, exprID)
}
