% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlcG2fun.R
\name{rlcG2fun}
\alias{rlcG2fun}
\title{Computes loglikelyhood of EVL models}
\usage{
rlcG2fun(temppars, tempparbounds, deckchoices, wins, losses)
}
\arguments{
\item{temppars}{temppars values}

\item{tempparbounds}{tempparbounds values}

\item{deckchoices}{deckchoices values}

\item{wins}{wins values}

\item{losses}{losses values}
}
\value{
log of temp choice probability
}
\description{
Computes loglikelyhood of EVL models
}
\examples{
# rlcG2fun(temppars,tempparbounds,deckchoices,wins,losses)
}
