% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalepars.R
\name{scalepars}
\alias{scalepars}
\title{scales the parameters for bounded optimals}
\usage{
scalepars(tpars, lb, ub)
}
\arguments{
\item{tpars}{enter tpars}

\item{lb}{lower bound}

\item{ub}{upper bound}
}
\description{
scales the parameters for bounded optimals
}
\examples{
# scalepars(tpars,lb,ub)
}
