% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscalepars.R
\name{unscalepars}
\alias{unscalepars}
\title{unscale the parameters for bounded optimals}
\usage{
unscalepars(tpars, lb, ub)
}
\arguments{
\item{tpars}{enter tpars}

\item{lb}{lower bound}

\item{ub}{upper bound}
}
\description{
unscale the parameters for bounded optimals
}
\examples{
# unscalepars(tpars,lb,ub)
}
